requirements.txt
adal==1.2.7
asttokens==3.0.0
azure-batch==6.0.0
azure-common==1.1.28
azure-core==1.32.0
azure-identity==1.19.0
azure-storage-blob==1.4.0
azure-storage-common==1.4.2
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
comm==0.2.2
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
dask==2025.1.0
debugpy==1.8.12
decorator==5.1.1
distributed==2025.1.0
executing==2.1.0
fonttools==4.55.3
fsspec==2024.12.0
idna==3.10
ipykernel==6.29.5
ipython==8.31.0
isodate==0.7.2
jedi==0.19.2
Jinja2==3.1.5
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
locket==1.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
msal==1.31.1
msal-extensions==1.2.0
msgpack==1.1.0
msrest==0.7.1
msrestazure==0.6.4.post1
nest-asyncio==1.6.0
numpy==2.2.1
oauthlib==3.2.2
packaging==24.2
pandas==2.2.3
parso==0.8.4
partd==1.4.2
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
portalocker==2.10.1
prompt_toolkit==3.0.48
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
requests==2.32.3
requests-oauthlib==2.0.0
scipy==1.15.0
six==1.17.0
sortedcontainers==2.4.0
stack-data==0.6.3
tblib==3.0.0
toolz==1.0.0
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
wcwidth==0.2.13
zict==3.0.0
